pyproject.toml
[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[project]
name = "gh-stars-organizer"
version = "0.1.0"
description = "Organize GitHub starred repositories into intelligent Star Lists with LLMs and embeddings."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "gh-stars-organizer contributors" }
]
dependencies = [
  "typer>=0.12.3",
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "rich>=13.7.1",
  "pyyaml>=6.0.1",
  "faiss-cpu>=1.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
]

[project.scripts]
gh-stars-organizer = "gh_stars_organizer.cli:app"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

